requirements.txt
# NoneBot2 核心框架
nonebot2[fastapi]>=2.4.4
nonebot-adapter-onebot>=2.4.6
nonebot-plugin-apscheduler>=0.5.0

# 异步工具
aiofiles
httpx>=0.27.0
anyio

# 数据处理
pydantic>=2.0.0
python-dateutil
requests

# 词云
jieba
wordcloud
numpy
Pillow

# 第三方 API
twikit==2.3.3
google-api-python-client  # YouTube
bilibili-api-python        # Bilibili

# 可选
atproto                    # Bluesky